Tag Search: move into expandable section for Channel Page
## Issue
2 search buttons deemed confusing. Moving it down to alleviate the issue for now.

## Changes
- Moved code into TagSearch component so we can place it in 2 places.
- Use context instead of props to pass in the "lift up" flag.
