+++ ui/component/claimListHeader/internal/tagSearch/style.scss
@@ -1,0 +1,50 @@
+@import '~ui/scss/init/breakpoints';
+
+$OPTIONS_MIN_WIDTH: 250px;
+
+.clh-tag-search {
+  .button-toggle {
+    border-radius: var(--border-radius);
+  }
+
+  .button-toggle--active {
+    border-radius: var(--border-radius) 0 0 var(--border-radius);
+    border-right: 1px solid var(--color-input-border);
+  }
+
+  input::placeholder {
+    font-size: var(--font-small);
+  }
+
+  .clh-tag-search__clear {
+    background: unset;
+    &:hover {
+      background: unset;
+    }
+  }
+
+  .clh-tag-search__clear--hidden {
+    display: none;
+  }
+}
+
+.clh-tag-search--standalone {
+  margin-left: var(--spacing-s);
+}
+
+.clh-tag-search__input_group {
+  display: flex;
+}
+
+.clh-tag-search__input {
+  width: $OPTIONS_MIN_WIDTH;
+}
+
+.clh-tag-search__input--standalone {
+  border-radius: 0 var(--border-radius) var(--border-radius) 0;
+  width: unset;
+}
+
+.clh-tag-search__input--hidden {
+  display: none;
+}
