+++ ui/component/claimListHeader/view.jsx
@@ -1,6 +1,7 @@
 // @flow
 import './style.scss';
 import AdditionalFilters from './internal/additionalFilters';
+import TagSearch from './internal/tagSearch/tagSearch';
 import * as CS from 'constants/claim_search';
 import * as ICONS from 'constants/icons';
 import * as SETTINGS from 'constants/settings';
@@ -9,7 +10,6 @@ import classnames from 'classnames';
 import React from 'react';
 import usePersistedState from 'effects/use-persisted-state';
 import usePersistentUserParam from 'effects/use-persistent-user-param';
-import { useIsLargeScreen } from 'effects/use-screensize';
 import { useHistory } from 'react-router';
 import { FormField } from 'component/common/form';
 import Button from 'component/button';
@@ -71,7 +71,6 @@ function ClaimListHeader(props: Props) {
   } = props;
 
   const filterCtx = React.useContext(ClaimSearchFilterContext);
-  const isLargeScreen = useIsLargeScreen();
   const { push, location } = useHistory();
   const { search } = location;
   const [expanded, setExpanded] = usePersistedState(`expanded-${location.pathname}`, false);
@@ -95,25 +94,16 @@ function ClaimListHeader(props: Props) {
     []
   );
 
-  const [tagSearchExpanded, setTagSearchExpanded] = React.useState(isLargeScreen);
-  const [tagSearchQuery, setTagSearchQuery] = React.useState(urlParams.get(CS.TAGS_KEY) || '');
-  const handleChangeDebounced = React.useCallback(
-    debounce((v) => handleChange({ key: CS.TAGS_KEY, value: v }), 500),
-    []
-  );
-
   const isFiltered = () =>
     Boolean(
       urlParams.get(CS.FRESH_KEY) ||
         urlParams.get(CS.CONTENT_KEY) ||
+        (!filterCtx?.liftUpTagSearch && urlParams.get(CS.TAGS_KEY)) ||
         urlParams.get(CS.DURATION_KEY) ||
         urlParams.get(CS.FEE_AMOUNT_KEY) ||
         urlParams.get(CS.LANGUAGE_KEY) ||
         filterCtx?.repost?.hideReposts
     );
-
-  // Pulled out of the collapsible group.
-  const isTagFiltered = urlParams.get(CS.TAGS_KEY);
 
   const languageValue = searchInLanguage
     ? languageParam === null
@@ -285,47 +275,7 @@ function ClaimListHeader(props: Props) {
               />
             )}
 
-            <div
-              className="clh-tag-search"
-              title={__('Multiple tags can be added by separating them with a comma.\nExample: sports,news,tv')}
-            >
-              <Button
-                icon={ICONS.TAG}
-                button="alt"
-                className={classnames('button-toggle', {
-                  'button-toggle--active': tagSearchExpanded,
-                  'button-toggle--custom': isTagFiltered,
-                })}
-                aria-label={__('Search tags')}
-                onClick={() => setTagSearchExpanded((prev) => !prev)}
-              />
-              <FormField
-                placeholder={__('Search tags')}
-                type="text"
-                className={classnames('clh-tag-search__input', {
-                  'clh-tag-search__input--hidden': !tagSearchExpanded,
-                })}
-                name="tag_query"
-                value={tagSearchQuery}
-                onChange={(e) => {
-                  setTagSearchQuery(e.target.value);
-                  handleChangeDebounced(e.target.value);
-                }}
-              />
-              <Button
-                icon={ICONS.REMOVE}
-                aria-label={__('Clear')}
-                button="alt"
-                className={classnames('clh-tag-search__clear', {
-                  'clh-tag-search__clear--hidden': !tagSearchExpanded || !tagSearchQuery,
-                })}
-                onClick={() => {
-                  setTagSearchQuery('');
-                  setTagSearchExpanded(false);
-                  handleChange({ key: CS.TAGS_KEY, value: '' });
-                }}
-              />
-            </div>
+            {filterCtx?.liftUpTagSearch && <TagSearch standalone urlParams={urlParams} handleChange={handleChange} />}
           </div>
         </div>
         {expanded && (
@@ -512,6 +462,8 @@ function ClaimListHeader(props: Props) {
                 </div>
               )}
             </div>
+
+            {!filterCtx?.liftUpTagSearch && <TagSearch urlParams={urlParams} handleChange={handleChange} />}
 
             {/* DURATIONS FIELD */}
             {showDuration && (
